grub: fix startup.nsh when built in bash shell
Remove extraneous backslashes in the statement written to startup.nsh.

Because it is enclosed in single quotes, the backslashes are not needed.
The `sh` shell will treat the extra backslashes as escape characters,
but `bash` does not. Further, `sh` does not require the backslashes to
be escaped, so removing the backslashes results in the correct behavior
on all shells.

Signed-off-by: David Lechner <dlechner@baylibre.com>