meta-synaptics:v4l2isp-daemon: add missing compilation flags
On 'synasdk-v4l2isp-daemon' recipe, it is linking against 'libdrm', but
the include path is missing from the recipe.

It seems that it was using the hosts header if they were available,
that's why we haven't seen this issue before building with crops.
Now, passing the correct include path, the build issues are solved.

Related-to: TOR-4075

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 92bce831620f7928122c96d603d610265df39366)