rac: remove -fcanon-prefix-map from CFLAGS
Build fails when compiling libz-sys with option -fcanon-prefix-map'.
This occurs because Rust build scripts compile C code using host GCC,
which might not support the -fcanon-prefix-map flag, that was introduced
in GCC 13+.
The flag was added to openembedded-core's bitbake.conf[1] specifically for
GCC 13+ cross-compilers, but gets inherited by host toolchain via CFLAGS
environment variable.
And we're seeing this issue since our crops build container has GCC
12.2.0, which is older than the required version.

To fix this issue, we remove -fcanon-prefix-map from CFLAGS.
This will have no impact to us, since we're still using -fdebug-prefix-map and
-fmacro-prefix-map, which provide equivalent functionality. Also
openembedded-core has removed this flag on another commit[2], that was
not ported to scarthgap.

[1] https://github.com/openembedded/openembedded-core/commit/abe97fa74b10c2c71d79bc3573d4edb509b111f5
[2] https://github.com/openembedded/openembedded-core/commit/ff73fa7ef7666a6dbe34f15515bc3ab6e574c5b0

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>