imx93frdm: Adapt wifi firmware to Torizon base layers.
The firmware-nxp-wifi recipe has been modified in the version of meta-imx
Torizon uses so that it is no longer buildable with the current bbappend
in meta-imx-frdm.  This is a temporary change to reset the packaging to
what is expected by the FRDM layer.  Once that layer is updated we can
drop this.

Related-to: TOR-4074

Signed-off-by: Drew Moseley <drew.moseley@toradex.com>