setup-environment-synaptics: make DL_DIR/SSTATE_DIR default values
This is similar to 655341c9f6089f7e9b7d2b14bd2e639ecf0745dc.

Originally our setup-environment-synaptics script was setting variables
DL_DIR and SSTATE_DIR in "local.conf" with a normal assignment (=) which
would override values previously set in "auto.conf". Doing this breaks
expectations of our CI scripts which set the same variables but the
settings don't get honored. To solve this, here we simply make the
assignments with the ?= operator allowing the settings in "auto.conf" to
take effect.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>