intel-corei7-64: fix fw_pritenv/fw_setenv on x86 machines
Currently u-boot-fw-utils is being set to grub-ota-fallback, which
installs the fw_printenv and fw_setenv into the torizon image. However,
this value is not being reflected:

$ bitbake-getvar -r u-boot-fw-utils --value PN libubootenv

This happends because the libuootenv from oe-core layer sets "PROVIDES =
u-boot-fw-utils", which has higher priority over the preferred provider
variable, selecting the wrong recipe.

Because of that, remove the preferred provider since x86 does not use
u-boot anyways and use the IMAGE_INSTALL to select the correct package.

Related-to: TOR-4047
Fixes: d372817 ("intel-corei7-64: add x86 support through meta-intel's intel-corei7-64")
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>