sl1680.inc: refactor synaptics meta files customization
Synaptics generates some meta files, emmc_part_list and emmc_image_list,
which we later customize to remove unused partitions, and add the
partition fo Torizon rootfs.

This customization conflicts with Synaptics image generation command,
where it expects a known list of partition names, and on Torizon we add
a partition with another name.

So, to avoid this conflict we first copy the meta files generated by
Synaptics into SYNAIMG folder, and then we customize it to our use. This
way, subsequent builds will only see the files generated by Synaptics
and this conflict is avoided.

Related-to: TOR-4087
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>