ostree: save torizon_bootargs_{l,r} from previous deployment
Replace previous patch which saved a variable "torizon_boot_args" with a
new patch that saves variables torizon_bootargs_l and torizon_bootargs_r
instead. The idea is to use the former for prepending and the latter for
appending to the kernel boot arguments (to be done on a separate commit
changing the boot script).

Also, take the opportunity to rebase the existing ostree patches.

Related-to: TCB-763
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>
(adapted from commit 6812726fd545534bb918f818dd09decc623fe21a)