meta-beagle:linux: enable Toshiba TC358768 MIPI DSI bridge
After bumping meta-ti, we started facing an issue on BeagleY AI builds,
where initramfs was failing to build due to it no finding the module for
TC358768.
Seems that after the bump, we this config was no longer set as module,
but as 'not set', causing this failure.
So now we add this config to our fragment, as a module, to fix the build
issue.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>